requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
PyMySQL==1.1.1
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
